(fix) replace remaining non standard str-related functions seperated portable mmap/unmap definitions and impl in own files, added define for use the fallback mmap impl
